pyproject.toml
[project]
name = "ss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "pydantic>=2.10.6",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
